requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
click==8.1.7
coverage==7.6.1
exceptiongroup==1.2.2
factory-boy==3.3.1
Faker==28.4.1
fastapi==0.114.0
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.9
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psutil==6.0.0
psycopg2-binary==2.9.9
pydantic==2.9.1
pydantic-core==2.23.3
pydantic-settings==2.5.2
pytest==8.3.3
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
tomli==2.0.1
typing-extensions==4.12.2
uvicorn==0.30.6
